replace the Integer Intervals to Long Intervals

+++ src/main/java/com/intervals/controller/IntervalsDTO/FactoryDTO.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 import java.util.List;
 
 public class FactoryDTO {
-    public static List<Integer> createGet(DigitsInterval interval){
+    public static List<Long> createGet(DigitsInterval interval){
         if (interval == null) return null;
         return Arrays.asList(interval.getStart(), interval.getEnded());
     }
